Split each class in a single file
