+++ firestarter/workflows/build_images/providers/registries/azure.py
@@ -1,0 +1,8 @@
+
+from .base import RegistryProvider
+
+
+class AzureOidcDockerRegistryAuth(RegistryProvider):
+
+    def get_registry_auth(self):
+        raise NotImplementedError('Azure OIDC not implemented yet')
