+++ firestarter/workflows/build_images/providers/registries/generic.py
@@ -1,0 +1,7 @@
+from .base import RegistryProvider
+
+
+class GenericDockerRegistryAuth(RegistryProvider):
+
+    def get_registry_auth(self):
+        raise NotImplementedError('Generic OIDC not implemented yet')
