+++ firestarter/workflows/build_images/providers/secrets/azure.py
@@ -1,0 +1,5 @@
+from .base import SecretProvider
+
+
+class AzureKeyVaultManager(SecretProvider):
+    pass
