+++ firestarter/workflows/build_images/providers/secrets/generic.py
@@ -1,0 +1,6 @@
+from .base import SecretProvider
+
+
+class GenericSecretManager(SecretProvider):
+    def get_secret(self):
+        return self.secret
